+++ packages/@baldeweg/ui/src/composables/__tests__/useGCPAuth.spec.js
@@ -1,0 +1,101 @@
+import { describe, it, expect, afterEach, vi } from 'vitest'
+import { useAuth } from '../useGCPAuth.js'
+
+const promise = () => {
+  return new Promise((resolve) => {
+    resolve()
+  })
+}
+
+vi.mock('firebase/app', () => {
+  const initializeApp = vi.fn()
+
+  return { initializeApp }
+})
+
+vi.mock('firebase/auth', () => {
+  const signInWithEmailAndPassword = () => {
+    return promise()
+  }
+  const signOut = () => {
+    return promise()
+  }
+  const updatePassword = () => {
+    return promise()
+  }
+  const getAuth = () => {
+    const setPersistence = vi.fn()
+    const currentUser = {}
+
+    return { setPersistence, currentUser }
+  }
+  const onAuthStateChanged = vi.fn()
+
+  return {
+    signInWithEmailAndPassword,
+    getAuth,
+    onAuthStateChanged,
+    signOut,
+    updatePassword,
+  }
+})
+
+const firebaseConfig = {
+  apiKey: 'key',
+  authDomain: 'localhost',
+}
+
+describe('GCPAuth', () => {
+  afterEach(() => {
+    vi.clearAllMocks()
+  })
+
+  const {
+    user,
+    isAuthenticated,
+    username,
+    password,
+    isLoggingIn,
+    isUpdatingPassword,
+    login,
+    logout,
+    updatePassword,
+  } = useAuth(firebaseConfig)
+
+  it('login', () => {
+    username.value = 'user'
+    password.value = 'pass'
+
+    login().then(() => {
+      expect(isAuthenticated.value).toBeTruthy()
+      expect(isLoggingIn.value).toBeFalsy()
+
+      expect(username.value).toBeNull()
+      expect(password.value).toBeNull()
+    })
+  })
+
+  it('logout', () => {
+    logout().then(() => {
+      expect(isAuthenticated.value).toBeFalsy()
+    })
+  })
+
+  it('user', () => {
+    login().then(() => {
+      expect(isAuthenticated.value).toBeTruthy()
+      expect(user.value).not.toBeNull()
+    })
+  })
+
+  it('update password', () => {
+    login().then(() => {
+      password.value = 'pass'
+
+      updatePassword().then(() => {
+        expect(password.value).toBeNull()
+        expect(isUpdatingPassword.value).toBeFalsy()
+      })
+    })
+  })
+})