+++ packages/@baldeweg/ui/src/mocks/resolvers/gcpAuth/mockSignInWithPassword.js
@@ -1,0 +1,15 @@
+export const mockSignInWithPassword = (req, res, ctx) => {
+  return res(
+    ctx.status(200),
+    ctx.json({
+      kind: 'identitytoolkit#VerifyPasswordResponse',
+      localId: 'id',
+      email: 'mail@localhost',
+      displayName: '',
+      idToken: 'token',
+      registered: true,
+      refreshToken: 'refresh_token',
+      expiresIn: '3600',
+    })
+  )
+}