+++ packages/@baldeweg/ui/src/mocks/resolvers/jwtAuth/mockPassword.js
@@ -1,0 +1,3 @@
+export const mockPassword = (req, res, ctx) => {
+  return res(ctx.status(200), ctx.json({}))
+}