+++ packages/@baldeweg/ui/src/mocks/resolvers/jwtAuth/mockToken.js
@@ -1,0 +1,9 @@
+export const mockToken = (req, res, ctx) => {
+  return res(
+    ctx.status(200),
+    ctx.json({
+      token: 'token',
+      refresh_token: 'refresh_token',
+    })
+  )
+}